build(pip): remove editable install of self
The "-e ." specifies an editable install of the current directory. This
is intended to cause the installation of the requirements to also
install the project itself, in a way that's suitable for development.

Unfortunately, the setup.py file is called to install this package. When
it's run, this file tries to import the current package. It does include
a shim so that it can import the source directly. However, during
import, it expects all of its dependencies to be installed already.

If pip read the requirements.txt file from top to bottom, and installed
packages in that order, this wouldn't be a problem: all of the
dependencies are declared first, and the installation of the package is
last.

However, the order in which pip installs requirements is not
deterministic:
https://github.com/pypa/pip/issues/2362#issuecomment-77464857

Normally, the dependencies are specified in setup.py; however, this
project relies on the guestfs package, which is currently not available
on PyPI, and only published to:
http://download.libguestfs.org/python

While pip and setuptools (the latter of which is what setup.py uses)
both provide ways to indicate a package located outside of PyPI, pip
requires this to be explicitly specified on the command line, and
ignores the setup.py specifying it:
https://github.com/pypa/pip/issues/4187#issuecomment-415067034

This all means that the current way to install this package, after this
commit, is to run the following two commands in the order given:
pip install -r requirements.txt
pip install .